+++ nucliadb_node/src/shards/mod.rs
@@ -23,7 +23,7 @@
 pub mod metadata;
 pub mod shard_reader;
 pub mod shard_writer;
-pub mod versions;
+pub mod versioning;
 
 // Alias for more readable imports
 pub use {shard_reader as reader, shard_writer as writer};
